[qtcontacts-sqlite] Remove duplicated effort from delta detection. Contributes to JB#32526
The delta-detection algorithm calculates the entire set of possible
permutations of (added, removed) details to determine the likely
set of (modified) details.

This commit ensures that the entire set of possible permutations for
each detail type is only calculated once, rather than once for
each potential modification.

Contributes to JB#32526