[qtcontacts-sqlite] Run tests with "make check"
Add a "check" target, so that tests can be run uninstalled, with "make
check". In the .pro files, define the TARGET variable before including
the common.pri file, which needs this variable to define the command to
be executed.

In order for the tests to load the plugin, it needs to be moved into a
"contacts" subdirectory; do this with the DESTDIR variable.