[qtcontacts-sqlite] Implement contact search filter query. Contributes to JB#47217
A common use-case is to search all contacts based on a simple search
string, where you want results to be ordered according to some
"likelihood" that the match is what the user is hoping to find.

This commit adds a new type of request, the search filter request,
which is implemented as a compound fetch request, which returns
results in order according to a built-in criteria of likelihood.

That criteria is:
- first name starts with
- last name starts with
- display label group starts with
- phone number starts with
- email or organization starts with
- any of the above details contains

(the substring provided as the search filter value).