[qtcontacts-sqlite] Support fine-grained change reporting. Contributes to JB#50551
Previously, contact sync operations required sync plugins to store
considerable amounts of data in out-of-band storage, in order to
correctly determine the local and remote change deltas.

This commit ensures that we store in the backend the precise
type of change which has occurred to each datum since the last
sync, so that these changes can be reported accurately to the sync
plugins.

The synchronization transaction API which is exposed to clients has
also been rewritten to allow changes to be made atomically for each
collection.

Finally, it improves the Collection support to allow arbitrary
metadata to be stored for any collection.  This allows e.g. an
addressbook ctag or synctoken to be stored in-band.