[qtcontacts-sqlite-extensions] Reimplement the TWCSA interface. Contributes to JB#50551
The two-way-contact-sync-adaptor interface exists to provide a simple
extension point for contact sync plugins with a well-defined contract
which, once implemented, will result in correct sync cycles.

As the sync transaction API in the backend has been significantly
changed (to support proper collection support with per-collection
transactions), the TWCSA interface and implementation needed to be
updated to use that updated sync transaction API.  This commit does
this.