Performance: improve performance of reads
When duplicates of a single contact are requested in a single read,
the order of detail ids will be identical in both duplicates, and
thus we don't need to store a hash of all previously seen detail ids,
but only the first detail id, in order to perform duplicate detection.

Also reduce branching in detail population, slightly.