[qtcontacts-sqlite] Time-optimise contact database queries. Contributes to JB#51444
Processing the contact details at the same time as the contact query
required both lists to be returned by sqlite in the same order, which
introduced a costly ORDER BY statement to the detail query.

This change caches the results in a hash table and then performs a
post-ordering step (which is fast because we already know the order). As
a result the detail query can return results in any order, avoiding the
need for sqlite to order the results, thus providing a performance bump.

The cost is additional memory overhead during the query.