[qtcontacts-sqlite] Auto-remove details from specific detail tables. Contributes to JB#50799
When a detail is removed, remove it from the detail-specific table,
instead of just the common detail table.

This was already implemented by createRemoveDetailsTrigger but this
trigger was accidentally omitted from the version 22 upgrade. So,
add the trigger to the common createStatements initialization, and
upgrade to version 23. When upgrading, remove all stale details found
in the detail-specific tables.