[qtcontacts-sqlite] Don't try loading sqlite icu extension by default. JB#57486
Seems I enabled this for all too hastily and whether the extension is
available or already built-in depends on the sqlite configuration.

The sqlite/ext/icu/README.txt tells how to create such an extension.
The code here, however, now fails on sailfish, though the sqlite itself is built
with ICU support. Executing "SELECT icu_load_collation('somelocale', 'localeCollation');"
succeeds the first time and second time it does a warning with a string
from the extension, proving also the extension is used.

This reverts commit 806cffe977f07dbcde80c0b25528734e9be0a8da.
This reverts commit 8c2f2960b8b5d57e21fa9120632f78c0fe1330de.