[qtcontacts-sqlite] Fix failing tst_qcontactmanager tests. JB#57252
0.3.11 / commit 6527885d8679 added timestamps to contact details
in order to know the latest changed avatar. That had a side-effect
of messing with QContactDetail comparisons due to timestamp differences
and broke some unit tests.

The change is slightly suspicious due to ending up affecting the
comparisons, and it's somewhat vague should that be taken into account
when doing it, but as the comparison is implemented in qtpim and
the timestamps are just extra fields here, it may not be that easy to
change.

Thus just adjusting the unit test specific expectations for now,
that should fine regardless of how QContactDetail::operator==() works.

tst_aggregation still failing, not yet sure what we should do there.